[examples][yolov2] rename yolov2/models -> yolov2/links
