+++ examples/detection/yolov2/yolov2/links/__init__.py
@@ -15,5 +15,5 @@
 from __future__ import print_function
 from __future__ import absolute_import
 
-from yolov2.models.yolov2 import YOLOv2, YOLOv2Predictor
-from yolov2.models.darknet19 import Darknet19, Darknet19Predictor
+from yolov2.links.yolov2 import YOLOv2, YOLOv2Predictor
+from yolov2.links.darknet19 import Darknet19, Darknet19Predictor
